Fix all cheats overriding SandboxAgentApplyDamageModel.CalculateDamage

+++ Patches/Combat/PartyDamageMultiplier.cs
@@ -10,7 +10,7 @@ namespace BannerlordCheats.Patches.Combat
     public static class PartyDamageMultiplier
     {
         [HarmonyPostfix]
-        public static void CalculateDamage(ref AttackInformation attackInformation, ref AttackCollisionData collisionData, WeaponComponentData weapon, ref int __result)
+        public static void CalculateDamage(ref AttackInformation attackInformation, ref AttackCollisionData collisionData, WeaponComponentData weapon, ref float __result)
         {
             if ((attackInformation.AttackerFormation?.Team?.IsPlayerTeam ?? false)
                 && !attackInformation.IsFriendlyFire
